Remove ConsulSchedulableLockProvider, add scheduled unlock for lockAtLeastFor for the ConsulLockProvider
